utils/weather_analyzer.py
from typing import Dict, Any
from datetime import datetime, timedelta

class WeatherAnalyzer:
    def __init__(self):
        # You might want to inject a database service here to fetch the data
        pass

    def get_daily_summary(self, city: str) -> Dict[str, Any]:
        # This is a placeholder implementation. In a real scenario, you'd fetch this data from your database.
        # For now, we'll return dummy data
        return {
            "city": city,
            "date": datetime.now().date().isoformat(),
            "average_temp": 25.5,
            "max_temp": 30.0,
            "min_temp": 20.0,
            "dominant_condition": "Clear"
        }

    def calculate_average_temp(self, temperatures: list[float]) -> float:
        return sum(temperatures) / len(temperatures) if temperatures else 0

    def get_dominant_condition(self, conditions: list[str]) -> str:
        return max(set(conditions), key=conditions.count) if conditions else "Unknown"